ulanzi-demos/usb/demo/pyproject.toml
[project]
name = "ulanzi-usb-demo"
version = "0.1.0"
description = "Ulanzi D200 USB dashboard demo – control D200 buttons via Python without UlanziStudio"
requires-python = ">=3.9"
dependencies = [
    "strmdck>=0.1.0rc1",
    "Pillow>=10.0.0",
    "psutil>=5.9.0",
]

[project.scripts]
ulanzi-usb-demo = "demo.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src/demo --cov-report=term-missing --cov-fail-under=80"

[tool.ruff]
line-length = 120

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
